+++ frontend/main/src/RarlaTodo/components/TodoList/index.scss
@@ -1,0 +1,19 @@
+ul, li {
+  margin: 0;
+  padding: 0;
+}
+
+li {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 5px;
+}
+
+button {
+  background-color: white;
+  border: none;
+}
+
+.completeTitle {
+  text-decoration: line-through;
+}
